requirements.txt
pygame-ce
raylib
pillow
lxml